String can be now used as input
Unit test is in the separate package, it is now test input as String.
All unit tests are not finished yet, but I've attached a comment in the
unit test file which shows you how to extract leaf in other programs. 
I add toString() methods to all nodes, I have a StringBuilder to collect
the data.